+++ ext_emconf.php
@@ -14,8 +14,8 @@ $EM_CONF[$_EXTKEY] = array(
 	'title' => 'Different HTTP statuscodes for shortcuts',
 	'description' => 'Restore the use of different HTTP statuscodes for shortcuts (307 redirect to first subpage, random subpage, parent page – 301 for redirect to explicitly selected page)',
 	'category' => 'misc',
-	'author' => 'Dan Untenzu',
-	'author_email' => 'untenzu@webit.de',
+	'author' => 'Dan Untenzu, Leonie Bitto [Netcreators]',
+	'author_email' => 'untenzu@webit.de, extensions@netcreators.nl',
 	'shy' => '',
 	'dependencies' => '',
 	'conflicts' => '',
@@ -29,7 +29,7 @@ $EM_CONF[$_EXTKEY] = array(
 	'clearCacheOnLoad' => 0,
 	'lockType' => '',
 	'author_company' => '',
-	'version' => '2.0.0',
+	'version' => '2.1.0',
 	'constraints' => array(
 		'depends' => array(
 			'typo3' => '6.2.0-7.6.99',
@@ -42,4 +42,3 @@ $EM_CONF[$_EXTKEY] = array(
 	),
 );
 
-?>